+++ packages/orq-rc/src/orq_ai_sdk/models/deploymentsop.py
@@ -520,7 +520,7 @@ DeploymentsRole = Literal[
 r"""The role of the prompt message"""
 
 
-Deployments2DeploymentsResponseType = Literal["file",]
+Deployments2DeploymentsType = Literal["file",]
 r"""The type of the content part. Always `file`."""
 
 
@@ -550,19 +550,19 @@ class Deployments2File(BaseModel):
 
 
 class Deployments2Deployments3TypedDict(TypedDict):
-    type: Deployments2DeploymentsResponseType
+    type: Deployments2DeploymentsType
     r"""The type of the content part. Always `file`."""
     file: Deployments2FileTypedDict
 
 
 class Deployments2Deployments3(BaseModel):
-    type: Deployments2DeploymentsResponseType
+    type: Deployments2DeploymentsType
     r"""The type of the content part. Always `file`."""
 
     file: Deployments2File
 
 
-Deployments2DeploymentsType = Literal["image_url",]
+Deployments2DeploymentsResponse200Type = Literal["image_url",]
 
 
 class Deployments2ImageURLTypedDict(TypedDict):
@@ -588,32 +588,32 @@ class Deployments2ImageURL(BaseModel):
 class Deployments2Deployments2TypedDict(TypedDict):
     r"""The image part of the prompt message. Only supported with vision models."""
 
-    type: Deployments2DeploymentsType
+    type: Deployments2DeploymentsResponse200Type
     image_url: Deployments2ImageURLTypedDict
 
 
 class Deployments2Deployments2(BaseModel):
     r"""The image part of the prompt message. Only supported with vision models."""
 
-    type: Deployments2DeploymentsType
+    type: Deployments2DeploymentsResponse200Type
 
     image_url: Deployments2ImageURL
 
 
-Deployments2DeploymentsResponse200Type = Literal["text",]
+Deployments2DeploymentsResponseType = Literal["text",]
 
 
 class Deployments2Deployments1TypedDict(TypedDict):
     r"""Text content part of a prompt message"""
 
-    type: Deployments2DeploymentsResponse200Type
+    type: Deployments2DeploymentsResponseType
     text: str
 
 
 class Deployments2Deployments1(BaseModel):
     r"""Text content part of a prompt message"""
 
-    type: Deployments2DeploymentsResponse200Type
+    type: Deployments2DeploymentsResponseType
 
     text: str