+++ packages/orq-rc/src/orq_ai_sdk/models/getbudgetop.py
@@ -161,5 +161,5 @@ class GetBudgetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-10-30T04:37:46.623Z")
+    updated: Optional[datetime] = parse_datetime("2025-10-31T10:51:55.522Z")
     r"""The date and time the resource was last updated"""