+++ packages/orq-rc/RELEASES.md
@@ -3749,3 +3749,13 @@ Based on:
 - [python v3.14.0-rc.70] packages/orq-rc
 ### Releases
 - [PyPI v3.14.0-rc.70] https://pypi.org/project/orq-ai-sdk/3.14.0-rc.70 - packages/orq-rc
+
+## 2025-10-31 10:53:44
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.643.3 (2.731.9) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.14.5-rc.1] packages/orq-rc
+### Releases
+- [PyPI v3.14.5-rc.1] https://pypi.org/project/orq-ai-sdk/3.14.5-rc.1 - packages/orq-rc